fix: Move addon code and remove app code
The first step in porting to v2 is to separate the addon code from the
test-app code. I've followed the standard of building a monorepo with
the addon code in `packages/ember-leaflet` and the (eventual) test-app
in `packages/test-app`. Additionally, the republishing code in `app/` is
no longer necessary. Finally, I've moved the addon's package.json into
the addon space, where it will be pared down.