chore(deps): Add ember-cli-htmlbars to addon to try and fix htmlbars test error
As discussed [on
discord](https://discord.com/channels/480462759797063690/973199754446131223/981925562769625150),
I've run into a problem with the various ember-leaflet components'
looking for htmlbars and not finding it. Even though I should not
technically have to include it as a dependency in the addon, the tests
crash because of it.

Additionally, the tests crash because of importing the components from
the addon and not the test-app, but I'll get to that problem later.